+++ tests/e2e/query_modules/CMakeLists.txt
@@ -4,4 +4,6 @@ endfunction()
 
 copy_query_modules_e2e_python_files(common.py)
 copy_query_modules_e2e_python_files(conftest.py)
-copy_query_modules_e2e_python_files(query_modules.py)
+copy_query_modules_e2e_python_files(convert_test.py)
+copy_query_modules_e2e_python_files(mgps_test.py)
+copy_query_modules_e2e_python_files(schema_test.py)
